requirements_image_downloader.txt
Pillow>=6.1.0
python-magic>=0.4.15
